requirements.txt
# Core ML/NLP Libraries
torch>=2.0.0
transformers>=4.35.0
peft>=0.7.0
accelerate>=0.24.0
bitsandbytes>=0.41.0
sentencepiece>=0.1.99
protobuf>=3.20.0

# RAG & Embeddings
faiss-cpu>=1.7.4
# faiss-gpu>=1.7.4  # Uncomment for GPU support
sentence-transformers>=2.2.2
langchain>=0.1.0
langchain-community>=0.0.10

# OCR & Document Processing
pytesseract>=0.3.10
Pillow>=10.0.0
pdf2image>=1.16.3
PyPDF2>=3.0.1
python-docx>=1.1.0

# Data Processing
pandas>=2.1.0
numpy>=1.24.0
scikit-learn>=1.3.0
nltk>=3.8.1
spacy>=3.7.0

# RLHF & Training
trl>=0.7.0
wandb>=0.16.0
tensorboard>=2.15.0

# Evaluation
rouge-score>=0.1.2
bert-score>=0.3.13
sacrebleu>=2.3.1

# Utilities
tqdm>=4.66.0
pyyaml>=6.0.1
python-dotenv>=1.0.0
requests>=2.31.0
aiohttp>=3.9.0

# Elasticsearch (optional, for alternative to FAISS)
# elasticsearch>=8.11.0

# Indic NLP (for Hindi processing)
indic-nlp-library>=0.81
indic-transliteration>=1.5.0

